requirements.txt
Python >= 3.5
SciPy >= 1.4.1
PyTorch >= 1.5.0
NumPy >= 1.18.1
scikit-learn >= 0.22.1
Matplotlib >= 3.1.1
chainer-chemistry